DeleteMessageTransactionWrapper и SaveMessageTransactionWrapper объединены в MessageTransactionWrapper. В DeleteMessageHandler удален self.
