+++ wfe-core/src/main/java/ru/runa/wfe/chat/logic/MessageTransactionWrapper.java
@@ -1,5 +1,6 @@
 package ru.runa.wfe.chat.logic;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +11,10 @@ import ru.runa.wfe.chat.ChatMessageFile;
 import ru.runa.wfe.chat.dao.ChatFileDao;
 import ru.runa.wfe.chat.dao.ChatMessageDao;
 import ru.runa.wfe.execution.dao.ProcessDao;
+import ru.runa.wfe.security.AuthorizationException;
 import ru.runa.wfe.user.Actor;
+import ru.runa.wfe.user.User;
+import ru.runa.wfe.user.logic.ExecutorLogic;
 
 /**
  * @author Alekseev Mikhail
@@ -18,7 +22,9 @@ import ru.runa.wfe.user.Actor;
  */
 @Transactional
 @Component
-public class SaveMessageTransactionWrapper {
+public class MessageTransactionWrapper {
+    @Autowired
+    private ExecutorLogic executorLogic;
     @Autowired
     private ChatMessageDao messageDao;
     @Autowired
@@ -38,5 +44,20 @@ public class SaveMessageTransactionWrapper {
     public ChatMessage save(ChatMessage message, Set<Actor> recipients, long processId) {
         message.setProcess(processDao.getNotNull(processId));
         return messageDao.save(message, recipients);
+    }
+
+    public List<ChatMessageFile> delete(User user, Long messageId) {
+        if (!executorLogic.isAdministrator(user)) {
+            throw new AuthorizationException("Allowed for admin only");
+        }
+        ChatMessage message = messageDao.getNotNull(messageId);
+        List<ChatMessageFile> files = fileDao.getByMessage(message);
+        messageDao.deleteMessageAndRecipient(messageId);
+        List<Long> ids = new ArrayList<>(files.size());
+        for (ChatMessageFile file : files) {
+            ids.add(file.getId());
+        }
+        fileDao.delete(ids);
+        return files;
     }
 }
